add sum

+++ src/flag_gems/sum.py
@@ -17,7 +17,7 @@ def sum_kernel_1(
     offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
     inp_ptrs = inp + offset
     mask = offset < M
-    inp_val = tl.load(inp_ptrs, mask=mask, other=0.0)
+    inp_val = tl.load(inp_ptrs, mask=mask, other=0.0).to(tl.float32)
     sum_val = tl.sum(inp_val)
     mid_ptr = mid + pid
     tl.store(mid_ptr, sum_val)
@@ -29,7 +29,7 @@ def sum_kernel_2(mid, out, mid_size, BLOCK_MID: tl.constexpr):
     offset = tl.arange(0, BLOCK_MID)
     mid_ptrs = mid + offset
     mask = offset < mid_size
-    mid_val = tl.load(mid_ptrs, mask=mask, other=0.0)
+    mid_val = tl.load(mid_ptrs, mask=mask, other=0.0).to(tl.float32)
     sum_val = tl.sum(mid_val)
     tl.store(out, sum_val)
 
@@ -73,7 +73,7 @@ def sum_kernel(
     mask1 = m_offset < M
     mask = m_offset[:, None] < M and n_offset[None, :] < N
     inp_ptrs = inp + offset
-    inp_vals = tl.load(inp_ptrs, mask=mask, other=0.0)
+    inp_vals = tl.load(inp_ptrs, mask=mask, other=0.0).to(tl.float32)
     result_index = tl.sum(inp_vals, axis=1)
 
     out_ptrs = out + offset_index
